# Faster fromListWord64