```
benchmarking Load Elias Fano/Load Elias Fano
time                 2.639 s    (2.426 s .. 2.883 s)
                     0.999 R²   (0.996 R² .. 1.000 R²)
mean                 2.623 s    (2.565 s .. 2.655 s)
std dev              54.77 ms   (4.030 ms .. 68.99 ms)
variance introduced by outliers: 19% (moderately inflated)
```